upgraded to Flask-JWT-Extended from Flask-JWT +logout, and blacklist

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity
 from models.item import ItemModel
 
 
@@ -23,9 +23,8 @@ class Item(Resource):
             return item.json()
         return {'message' : 'Item not found'}, 404
 
-
+    @jwt_required(fresh=True)
     def post(self, name):
-
         if ItemModel.find_by_name(name):
             return {'message' : "an item with {} name already exists".format(name)}, 400
             
@@ -52,8 +51,11 @@ class Item(Resource):
         item.save_to_db()
         return item.json()
 
-
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message' : 'Admin privelage required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -61,5 +63,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
-        return {'items' : [x.json() for x in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id: 
+            return {'items' : items}
+        return {
+            'items' : [item['name'] for item in items],
+            'message' : 'For more info log in'
+        }, 200
